apple: stabilize swift package ci on xcode 26.4
- bump Swift package manifests to tools 6.3 and refresh InlineMacUI resolved deps
- expand Swift package CI and prefer latest stable Xcode in macOS workflows
- move mocked auth/db-key tests off the real keychain and avoid shared DB/cache paths in package tests

+++ apple/InlineUI/Tests/InlineUITests/PlatformPhotoViewTests.swift
@@ -29,7 +29,7 @@ struct PlatformPhotoViewTests {
     let photoInfo = PhotoInfo(
       photo: photo,
       sizes: [
-        PhotoSize(photoId: 42, type: "s", width: 40, height: 30, size: Int64(strippedBytes.count), bytes: strippedBytes),
+        PhotoSize(photoId: 42, type: "s", width: 40, height: 30, size: strippedBytes.count, bytes: strippedBytes),
       ]
     )
